pyproject.toml
[tool.poetry]
name = "pyip"
version = "0.2.3"
description = ""
authors = ["Mikhail Isaev <admin@ismv.ru>"]
license = "AGPL-3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi-slim = "^0.111.0"
pydantic-settings = "^2.3.4"
maxminddb = "^2.6.2"
aiocache = "^0.12.2"
aiogram = "^3.8.0"
pyyaml = "^6.0.1"
uvicorn = {extras = ["standard"], version = "^0.30.1"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
httpx = "^0.27.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
